Yarn Audit
Yarn now has an audit command. So better than creating an npm lockfile
(which is very prone to failure) and then using `npm audit` we can use
`yarn audit` for yarn projects.

To make this as DRY as possible, we can create a superclass called
`JsAudit` to handle the logic we want to apply to both commands.

Currently `yarn audit` is broken by my currently implementaiton is
banking on the JSON outputs of each command being the same. When it's
working again I will test it annd if they are different we'll just have
to run some pre-processing logic to make them the same.

We will also only trigger these scans if the relevant lockfile is
present. This means that there is no longer a warning/failure for not
having a lockfile. This can be re-implemented later in a separate
module.

Configuration wise, anything that applies to JsAudit should apply to
both subclasses which is nice.

TOOD:
  - test on config patch