Scanners with raise do not fail builds
Scanners can fail for many reasons. There might be a bug in the scanner
or, a registry that it might rely on, such as a CVE database, might be
down. This would result in the scanner raising, and if the scanner is
required, Salus would then also fail.

This means that build pipelines might be heavily disrupted by external
dependencies which is not good. Furthermore, as we add more and more
scanners, the probability of this rises.

The status quo is that if a scanner raises, it counts as the scanner not
passing (as if it found an issue). Instead, we will count the scanner as
passed but still record the error. The build will pass overall and the
consumer of a Salus report can track the error.

This is probably a good setting to make configurable to the user and so
a future PR should allow for this to be toggled from Salus config files.

We add a test for this functionality which also adds a test for the
recording of errors to the Salus/Scan reports when a scanner raises.