Use Logs for Dev, Info for Reports
Previously, for a failed scanner, info data was presented to the
developer. This often has raw STDOUT from a scanner and other
information which makes the output that the developer needs to parse
very large and unreadable.

Instead we will use `log()` for all dev friendly output and it will be
output whenever a scanner fails. `info()` will be reserved for Salus
report consumers and verbose mode.